+++ NEWS.md
@@ -12,6 +12,8 @@ Bug fixes:
 - `read_cell_seg_data` recognizes and correctly reads inForm data 
   which uses comma as the decimal separator (#8).
 - Fix SpatialDistributionReport to work with more than 8 phenotypes (#6).
+- `count_touching_cells` ignores phenotype pairs of self to self with a warning
+  because the touching algorithm does not handle this case (#7).
 
 # phenoptr 0.1.6.1
 **2019-04-22**